+++ subethaedit/StylePreferences.m
@@ -506,11 +506,11 @@
 
 #pragma mark -
 #pragma mark TableView DataSource
-- (int)numberOfRowsInTableView:(NSTableView *)aTableView {
+- (NSInteger)numberOfRowsInTableView:(NSTableView *)aTableView {
     return [[I_currentSyntaxStyle allKeys] count];
 }
 
-- (id)tableView:(NSTableView *)aTableView objectValueForTableColumn:(NSTableColumn *)aTableColumn row:(int)aRow {
+- (id)tableView:(NSTableView *)aTableView objectValueForTableColumn:(NSTableColumn *)aTableColumn row:(NSInteger)aRow {
     BOOL darkBackground = ![[aTableColumn identifier]isEqualToString:@"light"];
     BOOL useDefault=[[[O_modePopUpButton selectedMode] defaultForKey:DocumentModeUseDefaultStylePreferenceKey] boolValue];
     NSString *key=[[I_currentSyntaxStyle allKeys] objectAtIndex:aRow];
@@ -555,7 +555,7 @@
 
 #pragma mark TableView Delegate
 
-- (BOOL)tableView:(NSTableView *)aTableView shouldSelectRow:(int)rowIndex {
+- (BOOL)tableView:(NSTableView *)aTableView shouldSelectRow:(NSInteger)rowIndex {
     BOOL useDefault=[[[O_modePopUpButton selectedMode] defaultForKey:DocumentModeUseDefaultStylePreferenceKey] boolValue];
     if (rowIndex==0 && useDefault) {
         return NO;
